requirements.txt
weaviate-client==4.18.0
sentence-transformers==5.1.2
python-dotenv==1.2.1
langchain==1.0.7
langchain-core==1.0.5
langchain-community==0.4.1
langchain-google-genai==3.0.3
langchain-classic==1.0.0
scikit-learn==1.7.2
numpy==2.2.6
joblib==1.5.2
pyvi==0.1.1
streamlit==1.51.0

